refactor(openomni): split skill manager into audit and io modules
Extract manager.ts (637 LOC) into 3 focused modules:
- manager.ts (242 LOC): SkillManager namespace operations only
- manager-audit.ts (269 LOC): audit pipeline, policy evaluation, Bus events
- manager-io.ts (129 LOC): registry/definition file I/O, serialization

Eliminates duplicate utility functions by importing from shared.ts.
