+++ packages/openomni/src/skill/manager.ts
@@ -1,8 +1,21 @@
-import { Policy, Operational, Skill } from "@openomni/protocol";
-import { Bus, Storage } from "@openomni/session";
-import { mkdir, readdir, rm } from "node:fs/promises";
-import { homedir } from "node:os";
-import { dirname, join } from "node:path";
+import { type Policy, Skill } from "@openomni/protocol";
+import {
+  type AuditState,
+  approveOperation,
+  beginOperation,
+  blockOperation,
+  operationInput,
+} from "./manager-audit";
+import {
+  globalManagerEntry,
+  readLocalEntries,
+  readRegistry,
+  removeSkillDirectory,
+  upsertRegistryEntry,
+  writeRegistry,
+  writeSkillDefinition,
+} from "./manager-io";
+import { resolveGlobalSkillsRoot, resolveLocalSkillsRoot, skillPath } from "./shared";
 
 type SkillAction =
   | "skill.install"
@@ -11,73 +24,7 @@ type SkillAction =
   | "skill.uninstall"
   | "skill.list";
 
-type AuditVisibility = "internal" | "llm_reason" | "user_audit";
-
-interface AuditBase {
-  readonly actionId: string;
-  readonly parentActionId?: string;
-  readonly visibility: AuditVisibility;
-  readonly timestamp: string;
-  readonly sequence: number;
-}
-
-interface AuditActionRequested extends AuditBase {
-  readonly type: "action_requested";
-  readonly actor: Record<string, unknown>;
-  readonly action: string;
-  readonly resource: string;
-  readonly input?: Record<string, unknown>;
-}
-
-interface AuditPolicyEvaluated extends AuditBase {
-  readonly type: "policy_evaluated";
-  readonly policyId: string;
-  readonly actor: Record<string, unknown>;
-  readonly action: string;
-  readonly resource: string;
-  readonly verdict: Policy.PolicyDecision["verdict"];
-  readonly reason: string;
-}
-
-interface AuditActionApproved extends AuditBase {
-  readonly type: "action_approved";
-  readonly policyId: string;
-  readonly actor: Record<string, unknown>;
-  readonly action: string;
-  readonly resource: string;
-  readonly verdict: "allow";
-  readonly reason: string;
-}
-
-interface AuditActionBlocked extends AuditBase {
-  readonly type: "action_blocked";
-  readonly policyId: string;
-  readonly actor: Record<string, unknown>;
-  readonly action: string;
-  readonly resource: string;
-  readonly verdict: Policy.PolicyDecision["verdict"];
-  readonly reason: string;
-}
-
-type AuditEvent =
-  | AuditActionRequested
-  | AuditPolicyEvaluated
-  | AuditActionApproved
-  | AuditActionBlocked;
-
-interface ErrorWithCode {
-  readonly code?: unknown;
-}
-
-interface AuditState {
-  readonly sessionId: string;
-  readonly actor: Record<string, unknown>;
-  readonly action: SkillAction;
-  readonly resource: string;
-  readonly input: Record<string, unknown>;
-  readonly parentActionId?: string;
-  readonly now: () => Date;
-}
+const DEFAULT_VERSION = "0.0.0";
 
 export interface SkillManagerRoots {
   readonly projectRoot?: string;
@@ -120,12 +67,6 @@ export interface SkillManagerEntry {
   readonly version?: string;
   readonly installedAt?: number;
 }
-
-const SKILL_FILE_NAME = "SKILL.md";
-const REGISTRY_FILE_NAME = "installed_skills.json";
-const AUDIT_VISIBILITY: AuditVisibility = "internal";
-const DEFAULT_VERSION = "0.0.0";
-const MAX_AUDIT_TEXT_LENGTH = 256;
 
 export namespace SkillManager {
   export async function install(
@@ -212,7 +153,7 @@ export namespace SkillManager {
       }
 
       await approveOperation(operation, "local skill uninstall approved");
-      await rm(dirname(path), { recursive: true, force: true });
+      await removeSkillDirectory(path);
       return { id, scope, enabled: false, path };
     }
 
@@ -229,7 +170,7 @@ export namespace SkillManager {
       registry.filter((entry) => entry.id !== id),
       options,
     );
-    await rm(dirname(path), { recursive: true, force: true });
+    await removeSkillDirectory(path);
 
     return { ...existing, scope: "global", enabled: false, path };
   }
@@ -286,306 +227,6 @@ async function updateGlobalRegistryEntry(
   return globalManagerEntry(next, options);
 }
 
-async function beginOperation(
-  options: SkillOperationOptions,
-  request: {
-    readonly action: SkillAction;
-    readonly resource: string;
-    readonly input: Record<string, unknown>;
-  },
-): Promise<AuditState> {
-  const sessionId = requireAuditSession(options);
-  const now = options.now ?? (() => new Date());
-  const requested = await appendAuditEvent(
-    sessionId,
-    "action_requested",
-    (base): AuditActionRequested => ({
-      type: "action_requested",
-      actor: options.actor,
-      action: request.action,
-      resource: request.resource,
-      input: request.input,
-      ...base,
-    }),
-    now,
-  );
-
-  const state: AuditState = {
-    sessionId,
-    actor: options.actor,
-    action: request.action,
-    resource: request.resource,
-    input: request.input,
-    parentActionId: requested.actionId,
-    now,
-  };
-  const result = Policy.evaluate(options.permission, {
-    action: request.action,
-    resource: request.resource,
-    input: request.input,
-    actor: options.actor,
-  });
-
-  await appendPolicyEvent(state, result);
-  if (result.action === "abort") {
-    await blockOperation(state, result.policyId, result.reason, canonicalVerdict(result));
-    throw new Error(
-      `Skill operation "${request.action}" on "${request.resource}" denied: ${result.reason}`,
-    );
-  }
-
-  return state;
-}
-
-function canonicalVerdict(result: Policy.EvaluationResult): Policy.PolicyDecision["verdict"] {
-  if (result.decision === "require_approval") return "pending";
-  return result.action === "continue" ? "allow" : "deny";
-}
-
-async function appendPolicyEvent(
-  state: AuditState,
-  result: Policy.EvaluationResult,
-): Promise<void> {
-  await appendAuditEvent(
-    state.sessionId,
-    "policy_evaluated",
-    (base): AuditPolicyEvaluated => ({
-      type: "policy_evaluated",
-      policyId: result.policyId,
-      actor: state.actor,
-      action: state.action,
-      resource: state.resource,
-      verdict: canonicalVerdict(result),
-      reason: result.reason,
-      ...base,
-    }),
-    state.now,
-    state.parentActionId,
-  );
-}
-
-async function approveOperation(state: AuditState, reason: string): Promise<void> {
-  await appendAuditEvent(
-    state.sessionId,
-    "action_approved",
-    (base): AuditActionApproved => ({
-      type: "action_approved",
-      policyId: "skill.manager",
-      actor: state.actor,
-      action: state.action,
-      resource: state.resource,
-      verdict: "allow",
-      reason,
-      ...base,
-    }),
-    state.now,
-    state.parentActionId,
-  );
-}
-
-async function blockOperation(
-  state: AuditState,
-  policyId: string,
-  reason: string,
-  verdict: AuditActionBlocked["verdict"] = "deny",
-): Promise<void> {
-  await appendAuditEvent(
-    state.sessionId,
-    "action_blocked",
-    (base): AuditActionBlocked => ({
-      type: "action_blocked",
-      policyId,
-      actor: state.actor,
-      action: state.action,
-      resource: state.resource,
-      verdict,
-      reason,
-      ...base,
-    }),
-    state.now,
-    state.parentActionId,
-  );
-}
-
-const auditSequences = new Map<string, number>();
-
-async function appendAuditEvent<T extends AuditEvent>(
-  sessionId: string,
-  eventType: T["type"],
-  event: (base: {
-    readonly actionId: string;
-    readonly parentActionId?: string;
-    readonly visibility: AuditVisibility;
-    readonly timestamp: string;
-    readonly sequence: number;
-  }) => T,
-  now: () => Date,
-  parentActionId?: string,
-): Promise<T> {
-  const prev = auditSequences.get(sessionId) ?? 0;
-  const sequence = prev + 1;
-  auditSequences.set(sessionId, sequence);
-  const row = event({
-    actionId: `${sessionId}:${eventType}:skill:${sequence}`,
-    ...(parentActionId !== undefined ? { parentActionId } : {}),
-    visibility: AUDIT_VISIBILITY,
-    timestamp: now().toISOString(),
-    sequence,
-  });
-
-  Bus.publish(Operational.Info, {
-    traceId: sessionId,
-    time: Date.now(),
-    sessionId,
-    component: "skill.manager",
-    msg: eventType,
-    context: { audit: row as unknown as Record<string, unknown> },
-  });
-  return row;
-}
-
-function requireAuditSession(options: SkillOperationOptions): string {
-  const sessionId = options.audit?.sessionId;
-  if (!sessionId) {
-    throw new Error("SkillManager operations require audit.sessionId");
-  }
-
-  const adapter = Storage.get();
-  if (!adapter.session.get(sessionId)) {
-    throw new Error(`Audit session "${sessionId}" not found for skill operation`);
-  }
-
-  return sessionId;
-}
-
-async function readRegistry(options: SkillManagerRoots): Promise<Skill.RegistryEntry[]> {
-  const path = resolveRegistryPath(options);
-  const file = Bun.file(path);
-  if (!(await file.exists())) {
-    return [];
-  }
-
-  let raw: unknown;
-  try {
-    raw = await file.json();
-  } catch (error) {
-    throw new Error(`Failed to read skill registry at ${path}: ${errorMessage(error)}`);
-  }
-
-  const parsed = Skill.RegistryEntry.array().safeParse(raw);
-  if (!parsed.success) {
-    throw new Error(`Invalid skill registry at ${path}: ${parsed.error.message}`);
-  }
-
-  return sortRegistryEntries(parsed.data);
-}
-
-async function writeRegistry(
-  entries: readonly Skill.RegistryEntry[],
-  options: SkillManagerRoots,
-): Promise<void> {
-  const parsed = Skill.RegistryEntry.array().safeParse(entries);
-  if (!parsed.success) {
-    throw new Error(`Invalid skill registry entries: ${parsed.error.message}`);
-  }
-
-  const path = resolveRegistryPath(options);
-  await mkdir(dirname(path), { recursive: true });
-  await Bun.write(path, `${JSON.stringify(sortRegistryEntries(parsed.data), null, 2)}\n`);
-}
-
-async function writeSkillDefinition(path: string, definition: Skill.Definition): Promise<void> {
-  await mkdir(dirname(path), { recursive: true });
-  await Bun.write(path, serializeSkillDefinition(definition));
-}
-
-function serializeSkillDefinition(definition: Skill.Definition): string {
-  const lines = [
-    "---",
-    `id: ${definition.id}`,
-    `name: ${definition.name}`,
-    `description: ${definition.description}`,
-    `layer: ${definition.layer}`,
-  ];
-
-  if (definition.useWhen !== undefined) {
-    lines.push(`useWhen: ${definition.useWhen}`);
-  }
-  if (definition.doNotUseWhen !== undefined) {
-    lines.push(`doNotUseWhen: ${definition.doNotUseWhen}`);
-  }
-  if (definition.finalChecklist !== undefined) {
-    lines.push("finalChecklist:");
-    for (const item of definition.finalChecklist) {
-      lines.push(`  - ${item}`);
-    }
-  }
-
-  const body = definition.promptFragment.trim() || definition.description;
-  return [...lines, "---", "", body, ""].join("\n");
-}
-
-function upsertRegistryEntry(
-  entries: readonly Skill.RegistryEntry[],
-  next: Skill.RegistryEntry,
-): Skill.RegistryEntry[] {
-  const without = entries.filter((entry) => entry.id !== next.id);
-  return sortRegistryEntries([...without, next]);
-}
-
-async function readLocalEntries(options: SkillManagerRoots): Promise<SkillManagerEntry[]> {
-  const root = resolveLocalSkillsRoot(options);
-  let entries: string[];
-  try {
-    entries = await readdir(root);
-  } catch (error) {
-    if (isEnoent(error)) {
-      return [];
-    }
-    throw error;
-  }
-
-  const skills: SkillManagerEntry[] = [];
-  for (const id of entries.sort((a, b) => a.localeCompare(b))) {
-    const path = skillPath(root, id);
-    if (await Bun.file(path).exists()) {
-      skills.push({ id, scope: "local", enabled: true, path });
-    }
-  }
-  return skills;
-}
-
-function globalManagerEntry(
-  entry: Skill.RegistryEntry,
-  options: SkillManagerRoots,
-): SkillManagerEntry {
-  return {
-    ...entry,
-    scope: "global",
-    path: skillPath(resolveGlobalSkillsRoot(options), entry.id),
-  };
-}
-
-function operationInput(input: {
-  readonly id?: string;
-  readonly scope?: Skill.Scope | "all";
-  readonly source?: string;
-  readonly enabled?: boolean;
-  readonly version?: string;
-}): Record<string, unknown> {
-  return {
-    ...(input.id !== undefined ? { id: truncateAuditText(input.id) } : {}),
-    ...(input.scope !== undefined ? { scope: input.scope } : {}),
-    ...(input.source !== undefined ? { source: truncateAuditText(input.source) } : {}),
-    ...(input.enabled !== undefined ? { enabled: input.enabled } : {}),
-    ...(input.version !== undefined ? { version: truncateAuditText(input.version) } : {}),
-  };
-}
-
-function truncateAuditText(value: string): string {
-  return value.length <= MAX_AUDIT_TEXT_LENGTH ? value : value.slice(0, MAX_AUDIT_TEXT_LENGTH);
-}
-
 function assertSafeSkillId(id: string): void {
   if (id !== "." && id !== ".." && /^[A-Za-z0-9._-]+$/.test(id)) {
     return;
@@ -593,45 +234,9 @@ function assertSafeSkillId(id: string): void {
   throw new Error(`Invalid skill id "${id}"`);
 }
 
-function resolveRegistryPath(options: SkillManagerRoots): string {
-  return (
-    options.registryPath ?? join(resolveHomeRoot(options.homeRoot), ".openomni", REGISTRY_FILE_NAME)
-  );
-}
-
-function resolveLocalSkillsRoot(options: SkillManagerRoots): string {
-  return (
-    options.localSkillsRoot ?? join(options.projectRoot ?? process.cwd(), ".openomni", "skills")
-  );
-}
-
-function resolveGlobalSkillsRoot(options: SkillManagerRoots): string {
-  return options.globalSkillsRoot ?? join(resolveHomeRoot(options.homeRoot), ".openomni", "skills");
-}
-
-function resolveHomeRoot(homeRoot: string | undefined): string {
-  return homeRoot ?? homedir();
-}
-
-function skillPath(root: string, id: string): string {
-  return join(root, id, SKILL_FILE_NAME);
-}
-
-function sortRegistryEntries(entries: readonly Skill.RegistryEntry[]): Skill.RegistryEntry[] {
-  return [...entries].sort((a, b) => a.id.localeCompare(b.id));
-}
-
 function sortManagerEntries(entries: readonly SkillManagerEntry[]): SkillManagerEntry[] {
   return [...entries].sort((a, b) => {
     const scopeComparison = a.scope.localeCompare(b.scope);
     return scopeComparison === 0 ? a.id.localeCompare(b.id) : scopeComparison;
   });
-}
-
-function isEnoent(error: unknown): boolean {
-  return typeof error === "object" && error !== null && (error as ErrorWithCode).code === "ENOENT";
-}
-
-function errorMessage(error: unknown): string {
-  return error instanceof Error ? error.message : String(error);
 }
